[refactor] 여행 디렉토링 moved
