+++ src/main/java/potatoes/server/chat/entity/Chat.java
@@ -14,7 +14,7 @@ import jakarta.persistence.OneToOne;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import potatoes.server.entity.Travel;
+import potatoes.server.travel.entity.Travel;
 import potatoes.server.user.entity.User;
 
 @Getter
