+++ src/main/java/potatoes/server/travel/entity/TravelUser.java
@@ -1,4 +1,4 @@
-package potatoes.server.entity;
+package potatoes.server.travel.entity;
 
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
