+++ src/main/java/potatoes/server/travel/repository/TravelPlanRepository.java
@@ -1,11 +1,11 @@
-package potatoes.server.repository;
+package potatoes.server.travel.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import potatoes.server.entity.Travel;
-import potatoes.server.entity.TravelPlan;
+import potatoes.server.travel.entity.Travel;
+import potatoes.server.travel.entity.TravelPlan;
 
 public interface TravelPlanRepository extends JpaRepository<TravelPlan, Long> {
