+++ src/main/java/potatoes/server/utils/annotation/AuthorizationArgumentResolver.java
@@ -1,6 +1,6 @@
 package potatoes.server.utils.annotation;
 
-import static potatoes.server.error.ErrorCode.*;
+import static potatoes.server.utils.error.ErrorCode.*;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -11,7 +11,7 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
-import potatoes.server.error.exception.WeGoException;
+import potatoes.server.utils.error.exception.WeGoException;
 import potatoes.server.utils.jwt.JwtTokenUtil;
 
 @Component
