+++ src/main/java/potatoes/server/utils/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package potatoes.server.error;
+package potatoes.server.utils.error;
 
 import java.util.ArrayList;
 import java.util.List;
